+++ packages/react-client/src/types/internal.ts
@@ -81,7 +81,7 @@ export type ScreenShareState = (
 export interface TrackManager {
   initialize: (deviceId?: string) => Promise<void>;
   stop: () => Promise<void>;
-  startStreaming: (startStreamingProps: StartStreamingProps) => Promise<string>;
+  startStreaming: (startStreamingProps?: StartStreamingProps) => Promise<string>;
   stopStreaming: () => Promise<void>;
   pauseStreaming: () => Promise<void>;
   resumeStreaming: () => Promise<void>;
