+++ packages/react-client/src/utils/track.ts
@@ -1,6 +1,6 @@
-import type { FishjamClient, TrackContext } from "@fishjam-cloud/ts-client";
+import type { Encoding, FishjamClient, TrackContext } from "@fishjam-cloud/ts-client";
 import type { PeerMetadata, TrackMetadata } from "../types/internal";
-import type { Track } from "../types/public";
+import type { SimulcastBandwidthLimits, StartStreamingProps, Track } from "../types/public";
 
 // In most cases, the track is identified by its remote track ID.
 // This ID comes from the ts-client `addTrack` method.
@@ -57,3 +57,24 @@ export function setupOnEndedCallback(
     }
   });
 }
+const getDisabledEncodings = (activeEncodings: Encoding[]) => {
+  const allEncodings: Encoding[] = ["l", "m", "h"];
+  return allEncodings.filter((encoding) => !activeEncodings.includes(encoding));
+};
+
+export const getConfigAndBandwidthFromProps = (
+  props: StartStreamingProps,
+  simulcastBandwidthLimits: SimulcastBandwidthLimits,
+) => {
+  if ("simulcast" in props) {
+    const config = {
+      enabled: true,
+      activeEncodings: props.simulcast,
+      disabledEncodings: getDisabledEncodings(props.simulcast),
+    };
+    const bandwidth = new Map<Encoding, number>(Object.entries(simulcastBandwidthLimits) as [Encoding, number][]);
+    return [bandwidth, config] as const;
+  }
+
+  return [props.maxBandwidth] as const;
+};
